refactor(todo/repository): use Long instead of Integer for JpaTodoRepository ID type

+++ src/main/java/com/semsoko/webstartbackend/todo/repository/JpaTodoRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 
 @Profile("dev-postgres")
 @Repository
-public interface JpaTodoRepository extends JpaRepository<TodoEntity, Integer>{
+public interface JpaTodoRepository extends JpaRepository<TodoEntity, Long>{
 
 }
